+++ src/lib/Scenes/Home/Components/Sales/index.tsx
@@ -3,6 +3,7 @@ import { SectionList, StyleSheet } from "react-native"
 import { createFragmentContainer, graphql } from "react-relay"
 import LotsByFollowedArtists from "./Components/LotsByFollowedArtists"
 import { SaleList } from "./Components/SaleList"
+import { ZeroState } from "./Components/ZeroState"
 
 class Sales extends React.Component<Props> {
   get data() {
@@ -18,6 +19,10 @@ class Sales extends React.Component<Props> {
   }
 
   render() {
+    if (this.props.viewer.sales.length === 0) {
+      return <ZeroState />
+    }
+
     const sections = [
       {
         data: [{ data: this.data.liveAuctions }],