+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Optimize the Home/ForYou view by re-enabling removal of clipped subviews - alloy
 * Brings back artist rails on the Home/ForYou view - alloy
 * Fix issue where unmounted Home/ForYou rails would be refreshed - chris & alloy
+* Fix payment request component status not updating after payment - alloy
 
 ### 1.4.0-beta.11